+++ src/shared/ui/Input/Input.tsx
@@ -1,7 +1,7 @@
 import { InputHTMLAttributes, forwardRef } from 'react';
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
-  inputSize: 'small' | 'medium' | 'large' | 'xsmall';
+  inputSize: 'xsmall' | 'small' | 'medium' | 'large';
   isError?: boolean;
 }
